[PROGRAMMERS] GPT 추천 5일차 입국심사 + 징검다리
- 이진 탐색 문제

+++ programmers/han/Level 2-3/java/solution/autoever/binarysearch/Solution73.java
@@ -18,7 +18,7 @@ public class Solution73 {
         long answer = right;
 
         while (left <= right) {
-            long mid = (left + right) / 2;
+            long mid = (long)(left + right) / 2;
             long cnt = 0;
 
             for (int time : times) {
